Dpi awareness (#109)
* dpi awareness test

* dpi scaling updates

* dpi scaling updates

* missing mouse inputs fix (I hope finally)

* finishing up the dpi branch...

+++ WickedEngine/wiBackLog.h
@@ -27,5 +27,5 @@ namespace wiBackLog
 
 	void setBackground(wiGraphics::Texture* texture);
 	void setFontSize(int value);
-	void setFontRowspacing(int value);
+	void setFontRowspacing(float value);
 };
