+++ WickedEngine/wiSpriteFont_BindLua.cpp
@@ -87,8 +87,8 @@ int wiSpriteFont_BindLua::SetPos(lua_State* L)
 		Vector_BindLua* param = Luna<Vector_BindLua>::lightcheck(L, 1);
 		if (param != nullptr)
 		{
-			font.params.posX = (int)XMVectorGetX(param->vector);
-			font.params.posY = (int)XMVectorGetY(param->vector);
+			font.params.posX = XMVectorGetX(param->vector);
+			font.params.posY = XMVectorGetY(param->vector);
 		}
 		else
 			wiLua::SError(L, "SetPos(Vector pos) argument is not a vector!");
@@ -105,8 +105,8 @@ int wiSpriteFont_BindLua::SetSpacing(lua_State* L)
 		Vector_BindLua* param = Luna<Vector_BindLua>::lightcheck(L, 1);
 		if (param != nullptr)
 		{
-			font.params.spacingX = (int)XMVectorGetX(param->vector);
-			font.params.spacingY = (int)XMVectorGetY(param->vector);
+			font.params.spacingX = XMVectorGetX(param->vector);
+			font.params.spacingY = XMVectorGetY(param->vector);
 		}
 		else
 			wiLua::SError(L, "SetSpacing(Vector spacing) argument is not a vector!");
